refactor: app config, view interface 제거
