+++ src/main/java/ladder/Application.java
@@ -1,9 +1,17 @@
 package ladder;
 
-import ladder.config.AppConfig;
+
+import ladder.controller.LadderController;
+import ladder.domain.ladder.generator.RandomBooleanGenerator;
+import ladder.view.InputView;
+import ladder.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        AppConfig.ladderController().run();
+        new LadderController(
+                new InputView(),
+                new OutputView(),
+                new RandomBooleanGenerator()
+        ).run();
     }
 }
